+++ Tests/OCTToxTests.m
@@ -782,10 +782,11 @@ void mocked_tox_self_get_public_key(const Tox *tox, uint8_t *public_key);
     self.tox.delegate = OCMProtocolMock(@protocol(OCTToxDelegate));
     expectBlock(self.tox.delegate);
 
+    callBlock();
+
     XCTestExpectation *expectation = [self expectationWithDescription:@"callback"];
 
-    dispatch_async(dispatch_get_global_queue(DISPATCH_QUEUE_PRIORITY_DEFAULT, 0), ^{
-        callBlock();
+    dispatch_after(dispatch_time(DISPATCH_TIME_NOW, 0.1 * NSEC_PER_SEC), dispatch_get_main_queue(), ^{
         [expectation fulfill];
     });